% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_funs.R
\name{api_call}
\alias{api_call}
\title{Create API calls}
\usage{
api_call(
  url = NULL,
  start_date = NULL,
  end_date = NULL,
  treat_count = 100,
  profile_count = 50
)
}
\arguments{
\item{url}{The base url}

\item{start_date}{The date of the first treatment to be retrieved}

\item{end_date}{The date of the last treatment to be retrieved}

\item{treat_count}{The maximum number of treatments to be retrieved}

\item{profile_count}{The maximum number of profiles to be retrieved}
}
\value{

}
\description{
Create API calls
}
