% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{bind_insulin}
\alias{bind_insulin}
\title{Combine insulin basal and bolus}
\usage{
bind_insulin(basal_wide, bolus_wide)
}
\arguments{
\item{basal_wide}{}

\item{bolus_wide}{}
}
\description{
Combine insulin basal and bolus
}
