% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_funs.R
\name{bind_profiles}
\alias{bind_profiles}
\title{Tidy and combine a list of profiles json objects}
\usage{
bind_profiles(ls_profiles)
}
\arguments{
\item{ls_profiles}{}
}
\description{
Tidy and combine a list of profiles json objects
}
