% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_funs.R
\name{bind_treats}
\alias{bind_treats}
\title{Tidy and combine a list of treatment json objects}
\usage{
bind_treats(ls_treats)
}
\arguments{
\item{ls_treats}{}
}
\description{
Tidy and combine a list of treatment json objects
}
