% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{calc_basal}
\alias{calc_basal}
\title{Calculate insulin basal rates}
\usage{
calc_basal(dat_profile, dat_treats)
}
\arguments{
\item{dat_profile}{}

\item{dat_treats}{}
}
\description{
Calculate insulin basal rates
}
