% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_funs.R
\name{calc_insulin}
\alias{calc_insulin}
\title{Extract data from Nightscout and prepare summary of insulin use}
\usage{
calc_insulin(
  url = NULL,
  start_date = NULL,
  end_date = NULL,
  treat_count = 100,
  profile_count = 50
)
}
\arguments{
\item{url}{}

\item{start_date}{}

\item{end_date}{}

\item{treat_count}{}

\item{profile_count}{}
}
\value{

}
\description{
Extract data from Nightscout and prepare summary of insulin use
}
