% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requests.R
\name{find_between}
\alias{find_between}
\title{Convenience function for creating request query}
\usage{
find_between(x = "dateString", lower = NULL, upper = NULL, inc.bounds = TRUE)
}
\arguments{
\item{x}{An orderable vector, i.e., those with relevant methods for `<=`, usually a date field.}

\item{lower}{Lower range bound of length 1. Will be coerced to character if not already.}

\item{upper}{Upper range bound of length 1. Will be coerced to character if not already.}

\item{inc.bounds}{`TRUE` (default) means inclusive bounds, i.e., [lower,upper]. `FALSE` means exclusive bounds, i.e., (lower,upper).}
}
\value{
A list of maximum length 2, with expressions to use in later functions like `make_request`.
}
\description{
Convenience function for creating request query
}
