% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requests.R
\name{make_requests}
\alias{make_requests}
\title{Create multiple API request to retrieve data from Nightscout}
\usage{
make_requests(
  url = NULL,
  api_secret = NULL,
  schemas = c("entries", "treatments", "profile"),
  spec = "sgv",
  date_first = NULL,
  date_last = NULL,
  count = NULL
)
}
\arguments{
\item{url}{Base URL for accessing Nightscout of type `https://YOUR-SITE.com`. See https://nightscout-test.readthedocs.io/en/latest/Nightscout/EN/Technical%20info/api.html for further details. `ap/i/v1` is automatically added to the end of the base url when creating the request.}

\item{api_secret}{API secret or token as entered when accessing the base URL via a web browser. This API secret is hashed using Secure Hash Algorithm 1 prior to appending to header of API request.}

\item{schemas}{Vector of schema to be requested. For CGM data use `entries`, for Care Portal Treatments use `treatments`, for Treatment Profiles use `profiles`. and for server status and settings use `status`.
Relevant for schema `entries` only. entry id, such as 55cf81bc436037528ec75fa5 or a type filter such as sgv, mbg, etc. Default is 'sgv' which returns sensor glucose values (SGV).}

\item{date_first}{First date for `entries` and `treatments` schemas.}

\item{count}{By default the `entries` and `treatments` schemas limit results to the the most recent 10 values from the last 2 days. You can get many more results by setting this to a higher number. Can set to `Inf` to get all results between two dates defined in `find`. Not valid and ignored for `profile` and `status` schemas.}

\item{date_list}{Last date for `entries` and `treatments` schemas.}
}
\value{
A list of HTTP responses: a list of S3 lists with class httr2_request.
}
\description{
Create multiple API request to retrieve data from Nightscout
}
\examples{
\dontrun{
make_requests(
  url = "https://YOUR-SITE.com", 
  api_secret = "password",
  date_first = "2024-01-01", 
  date_last = Sys.Date())
}
 

}
