% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requests.R
\name{save_requests}
\alias{save_requests}
\title{Retrieve and save Nightscout data}
\usage{
save_requests(path = "", suffix = "", ...)
}
\arguments{
\item{path}{A folder location where retrieved data is to be saved.}

\item{suffix}{By default files are saved according to their schema (e.g. entries.Rds), a suffix may be applied to the filename.}

\item{...}{Additional arguments to pass to `make_requests`. Including base `url` and `api_secret`.}
}
\description{
Retrieve and save Nightscout data
}
