% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_funs.R
\name{tidy_treats}
\alias{tidy_treats}
\title{Tidy single nightscout treatments json object to data.table}
\usage{
tidy_treats(dat)
}
\arguments{
\item{dat}{}
}
\value{

}
\description{
Tidy single nightscout treatments json object to data.table
}
