% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_funs.R
\name{api_run}
\alias{api_run}
\title{Run API calls and retrieve data}
\usage{
api_run(api)
}
\arguments{
\item{api}{the url/api call of the page to retrieve}
}
\value{

}
\description{
Run API calls and retrieve data
}
