% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_funs.R
\name{tidy_treats}
\alias{tidy_treats}
\title{Tidy single nightscout treatments json object to data.table}
\usage{
tidy_treats(x)
}
\arguments{
\item{x}{json object of schema treatments}
}
\description{
Tidy single nightscout treatments json object to data.table
}
